requirements.txt
# pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124

ipdb==0.13.13
ipython==8.38.0

matplotlib==3.10.8
matplotlib-inline==0.2.1
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.15.3

tifffile==2025.5.10
tqdm==4.67.1

tensorrt-cu12
onnx==1.19.1
nvidia-modelopt[all]